backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.1
dj-database-url>=2.2
django-cors-headers>=4.4
